Feature Flag
Sets up feature flags to be more self contained.

This setup self wraps a Provider (context) around the Flags component making it easier to import and use.
